LayoutTests/imported/w3c/web-platform-tests/html/dom/render-blocking/parser-inserted-module-script-expected.txt

1

PASS Rendering is blocked before render-blocking resources are loaded
PASS Parser-inserted render-blocking module script is evaluated
